# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DEPOT API",
        "contact": {
            "name": "oussema khlifi",
            "email": "khoussemaa@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/providers/all": {
            "get": {
                "tags": [
                    "Providers"
                ],
                "summary": "Return list of providers",
                "operationId": "b5bc66ffd8c3962084ba705e324da25f",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProviderSchema"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/providers/create": {
            "post": {
                "tags": [
                    "Providers"
                ],
                "summary": "Provider store",
                "operationId": "6e0b2446ef80f434923d0a13793f2bff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "fax": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "mat": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "country": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProviderSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/providers/provider/{id}": {
            "get": {
                "tags": [
                    "Providers"
                ],
                "summary": "Provider show",
                "operationId": "d5e62b16fe80462f9e86c038cca7fb52",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProviderSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/providers/update/{id}": {
            "put": {
                "tags": [
                    "Providers"
                ],
                "summary": "Provider update",
                "operationId": "ac8f3d6b1c8f482739d1043ebce53e11",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "fax": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "mat": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "country": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProviderSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/providers/delete/{id}": {
            "delete": {
                "tags": [
                    "Providers"
                ],
                "summary": "Provider delete",
                "operationId": "06f398398e29649226a9f3c6e5c1cb3b",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ProviderSchema": {
                "title": "Provider Model",
                "description": "Provider Model",
                "properties": {
                    "name": {
                        "description": "provider name",
                        "type": "string",
                        "example": "samsung corp"
                    },
                    "email": {
                        "description": "provider email",
                        "type": "string",
                        "example": "contact@samsung.com"
                    },
                    "fax": {
                        "description": "provider fax",
                        "type": "string",
                        "example": "+216 74 523 658"
                    },
                    "phone": {
                        "description": "provider phone",
                        "type": "string",
                        "example": "+216 55 574 426"
                    },
                    "mat": {
                        "description": "provider registration number",
                        "type": "string",
                        "example": "Az258xfvbz"
                    },
                    "address": {
                        "description": "provider location",
                        "type": "string",
                        "example": "Sfax, route gremda km2"
                    },
                    "country": {
                        "description": "provider country",
                        "type": "string",
                        "example": "tunisia"
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "query",
                "description": "depot id",
                "schema": {
                    "type": "number"
                }
            }
        }
    }
}